+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-
+- `ReactEventEmmitter` and `ReactEventEmitterBuilder` binding for ReactPHP event loop
+- Idle worker threshold (I/O inactivity ms) to allow better control of idle workers
 
 [Unreleased]: https://github.com/ecomdev/reactive-socket/compare/4b825dc642cb6eb9a060e54bf8d69288fbee4904...HEAD